feat: added rate limiter
added rate limter for all requests

BREAKING CHANGE: Rate limter and display modification
