+++ static/api-interface.js
@@ -11,8 +11,10 @@ $(document).ready(function () {
             url: "/api/v1/fr-potato-api-input", type: "POST", // You can use GET or POST based on your API
             contentType: "application/json", data: JSON.stringify(filterData), success: function (data) {
                 dataTable.rows.add(data).draw();
-            }, error: function (error) {
-                console.error("Error:", error);
+            }, error: function (xhr, status, error) {
+                const {error: errorMessage, status: errorStatus} = $.parseJSON(xhr.responseText);
+                $("#errorAlert").html("<strong>Error " + errorStatus + ":</strong> " + errorMessage).show();
+                console.error("Error:", xhr);
             },
         });
     }
